requirements.txt
langchain==0.1.0
openai==1.3.0
python-dotenv==1.0.0
supabase==2.0.0
pypdf==3.17.1
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
pyyaml==6.0.1
python-magic==0.4.27
tiktoken==0.5.1 